Update migrations to be fully backwards compatible
Migrations now use a simple helper that creates a dummy model entry for the
specified user in the models dict of all migrations. This way we don't rely
on knowing what AUTH_USER_MODEL (if any) a project has specified.

Do not confuse this setting with the LAZYSIGNUP_USER_MODEL.